parser_worker/requirements.txt
pika
psycopg2-binary
python-dotenv
boto3